+++ test/my_package_test.exs
@@ -1,8 +1,0 @@
-defmodule MyPackageTest do
-  use ExUnit.Case
-  doctest MyPackage
-
-  test "example/1 doubles integers" do
-    assert MyPackage.example(3) == 6
-  end
-end