+++ test/safe_atom_test.exs
@@ -1,0 +1,8 @@
+defmodule SafeAtomTest do
+  use ExUnit.Case
+  doctest SafeAtom
+
+  test "example/1 doubles integers" do
+    assert SafeAtom.example(3) == 6
+  end
+end